Add sorting to market pairs table
Slight refactor for token metrics index page.
